# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for NeuralService Application",
        "title": "NeuralService API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/",
    "paths": {
        "/calls": {
            "get": {
                "description": "Get call information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "GetCall",
                "operationId": "get-call",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "model",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ServiceCall"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "description": "Store call information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "StoreCall",
                "operationId": "store-call",
                "parameters": [
                    {
                        "description": "call info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ServiceCall"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ServiceCall": {
            "type": "object",
            "properties": {
                "call_time": {
                    "type": "string"
                },
                "cost": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "model_id": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}